+++ internal/validation/validation.go
@@ -39,6 +39,7 @@ type context struct {
 	usedVars             map[*ast.OperationDefinition]varSet
 	fieldMap             map[*ast.Field]fieldInfo
 	overlapValidated     map[selectionPair]bool
+	allowDeprecatedUsage bool
 	maxDepth             int
 	overlapPairLimit     int
 	overlapPairsObserved int
@@ -57,26 +58,34 @@ func (c *context) addErrMultiLoc(locs []errors.Location, rule string, format str
 	})
 }
 
+func (c *context) addDeprecatedErr(loc errors.Location, format string, a ...any) {
+	if c.allowDeprecatedUsage {
+		return
+	}
+	c.addErr(loc, "NoDeprecatedCustomRule", format, a...)
+}
+
 type opContext struct {
 	*context
 	ops []*ast.OperationDefinition
 }
 
-func newContext(s *ast.Schema, doc *ast.ExecutableDefinition, maxDepth int, overlapPairLimit int) *context {
+func newContext(s *ast.Schema, doc *ast.ExecutableDefinition, maxDepth int, overlapPairLimit int, allowDeprecatedUsage bool) *context {
 	return &context{
-		schema:           s,
-		doc:              doc,
-		opErrs:           make(map[*ast.OperationDefinition][]*errors.QueryError),
-		usedVars:         make(map[*ast.OperationDefinition]varSet),
-		fieldMap:         make(map[*ast.Field]fieldInfo),
-		overlapValidated: make(map[selectionPair]bool),
-		maxDepth:         maxDepth,
-		overlapPairLimit: overlapPairLimit,
+		schema:               s,
+		doc:                  doc,
+		opErrs:               make(map[*ast.OperationDefinition][]*errors.QueryError),
+		usedVars:             make(map[*ast.OperationDefinition]varSet),
+		fieldMap:             make(map[*ast.Field]fieldInfo),
+		overlapValidated:     make(map[selectionPair]bool),
+		allowDeprecatedUsage: allowDeprecatedUsage,
+		maxDepth:             maxDepth,
+		overlapPairLimit:     overlapPairLimit,
 	}
 }
 
-func Validate(s *ast.Schema, doc *ast.ExecutableDefinition, variables map[string]any, maxDepth int, overlapPairLimit int) []*errors.QueryError {
-	c := newContext(s, doc, maxDepth, overlapPairLimit)
+func Validate(s *ast.Schema, doc *ast.ExecutableDefinition, variables map[string]any, maxDepth int, overlapPairLimit int, allowDeprecatedUsage bool) []*errors.QueryError {
+	c := newContext(s, doc, maxDepth, overlapPairLimit, allowDeprecatedUsage)
 
 	opNames := make(nameSet, len(doc.Operations))
 	fragUsedBy := make(map[*ast.FragmentDefinition][]*ast.OperationDefinition)
@@ -629,7 +638,7 @@ func validateSelection(c *opContext, sel ast.Selection, t ast.NamedType) {
 		validateArgumentLiterals(c, sel.Arguments)
 		if f != nil {
 			if reason, ok := deprecatedReason(f.Directives); ok && t != nil {
-				c.addErr(sel.Name.Loc, "NoDeprecatedCustomRule", "The field %s.%s is deprecated. %s", t.TypeName(), fieldName, reason)
+				c.addDeprecatedErr(sel.Name.Loc, "The field %s.%s is deprecated. %s", t.TypeName(), fieldName, reason)
 			}
 
 			validateArgumentTypes(c, sel.Arguments, f.Arguments, sel.Alias.Loc,
@@ -644,7 +653,7 @@ func validateSelection(c *opContext, sel ast.Selection, t ast.NamedType) {
 						continue
 					}
 					if reason, ok := deprecatedReason(argDecl.Directives); ok {
-						c.addErr(selArg.Name.Loc, "NoDeprecatedCustomRule", "Field %q argument %q is deprecated. %s", t.TypeName()+"."+fieldName, selArg.Name.Name, reason)
+						c.addDeprecatedErr(selArg.Name.Loc, "Field %q argument %q is deprecated. %s", t.TypeName()+"."+fieldName, selArg.Name.Name, reason)
 					}
 				}
 
@@ -659,7 +668,7 @@ func validateSelection(c *opContext, sel ast.Selection, t ast.NamedType) {
 							continue
 						}
 						if reason, ok := deprecatedReason(argDecl.Directives); ok {
-							c.addErr(selArg.Name.Loc, "NoDeprecatedCustomRule", "Directive %q argument %q is deprecated. %s", "@"+directive.Name.Name, selArg.Name.Name, reason)
+							c.addDeprecatedErr(selArg.Name.Loc, "Directive %q argument %q is deprecated. %s", "@"+directive.Name.Name, selArg.Name.Name, reason)
 						}
 					}
 				}
@@ -1079,7 +1088,7 @@ func validateDirectives(c *opContext, loc string, directives ast.DirectiveList) 
 					continue
 				}
 				if reason, ok := deprecatedReason(argDecl.Directives); ok {
-					c.addErr(selArg.Name.Loc, "NoDeprecatedCustomRule", "Directive %q argument %q is deprecated. %s", "@"+dirName, selArg.Name.Name, reason)
+					c.addDeprecatedErr(selArg.Name.Loc, "Directive %q argument %q is deprecated. %s", "@"+dirName, selArg.Name.Name, reason)
 				}
 			}
 		}
@@ -1268,7 +1277,7 @@ func validateValueType(c *opContext, v ast.Value, t ast.Type) (bool, errors.Loca
 				continue
 			}
 			if depReason, deprecated := deprecatedReason(option.Directives); deprecated {
-				c.addErr(lit.Location(), "NoDeprecatedCustomRule", "The enum value %q is deprecated. %s", enumType.Name+"."+option.EnumValue, depReason)
+				c.addDeprecatedErr(lit.Location(), "The enum value %q is deprecated. %s", enumType.Name+"."+option.EnumValue, depReason)
 			}
 			break
 		}
@@ -1303,7 +1312,7 @@ func validateValueType(c *opContext, v ast.Value, t ast.Type) (bool, errors.Loca
 				return false, f.Name.Loc, fmt.Sprintf("Field %q is not defined by type %q.%s", name, t.Name, suggestion)
 			}
 			if depReason, deprecated := deprecatedReason(iv.Directives); deprecated {
-				c.addErr(f.Name.Loc, "NoDeprecatedCustomRule", "The input field %s.%s is deprecated. %s", t.Name, iv.Name.Name, depReason)
+				c.addDeprecatedErr(f.Name.Loc, "The input field %s.%s is deprecated. %s", t.Name, iv.Name.Name, depReason)
 			}
 			if ok, errLoc, reason := validateValueType(c, f.Value, iv.Type); !ok {
 				return false, errLoc, reason